requirements.txt
attrs==20.2.0
aws-cdk.assets==1.69.0
aws-cdk.aws-applicationautoscaling==1.69.0
aws-cdk.aws-autoscaling-common==1.69.0
aws-cdk.aws-cloudwatch==1.71.0
aws-cdk.aws-codeguruprofiler==1.69.0
aws-cdk.aws-ec2==1.69.0
aws-cdk.aws-efs==1.69.0
aws-cdk.aws-events==1.71.0
aws-cdk.aws-iam==1.71.0
aws-cdk.aws-kinesis==1.69.0
aws-cdk.aws-kinesisanalytics==1.69.0
aws-cdk.aws-kinesisfirehose==1.69.0
aws-cdk.aws-kms==1.71.0
aws-cdk.aws-lambda==1.69.0
aws-cdk.aws-logs==1.69.0
aws-cdk.aws-s3==1.69.0
aws-cdk.aws-s3-assets==1.69.0
aws-cdk.aws-sns==1.71.0
aws-cdk.aws-sqs==1.71.0
aws-cdk.aws-ssm==1.69.0
aws-cdk.cloud-assembly-schema==1.71.0
aws-cdk.core==1.71.0
aws-cdk.cx-api==1.71.0
aws-cdk.region-info==1.71.0
bleach==3.2.1
boto3==1.16.0
botocore==1.19.0
cattrs==1.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
constructs==3.2.11
cycler==0.10.0
docutils==0.16
idna==2.10
jmespath==0.10.0
jsii==1.14.1
keyring==21.4.0
kiwisolver==1.2.0
matplot==0.1.9
matplotlib==3.3.2
numpy==1.19.2
packaging==20.4
Pillow==8.0.1
pkginfo==1.6.1
publication==0.0.3
Pygments==2.7.2
pyloco==0.0.139
pyparsing==2.4.7
python-dateutil==2.8.1
readme-renderer==28.0
requests==2.24.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
s3transfer==0.3.3
SimpleWebSocketServer==0.1.1
six==1.15.0
tqdm==4.51.0
twine==3.2.0
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.25.11
ushlex==0.99.1
webencodings==0.5.1
websocket-client==0.57.0
